# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoCRM API",
    "description": "Autotherm project lifecycle system. Money is integer minor units with an explicit currency; quantities are decimal strings. PATCH: omit a field to keep it, send null to clear it.",
    "license": {
      "name": ""
    },
    "version": "1"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/admin/status": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_status",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatus"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/admin/jobs": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_list_jobs",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "\"failed\" (default) or \"pending\"",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_Job"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/admin/jobs/{id}/retry": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_retry_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Re-queued"
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/admin/fx/fetch": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "admin_fetch_fx",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FxFetch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobQueued"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/admin/run/{kind}": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Runs a periodic job now instead of waiting for the scheduler (\"send the nudges now\").",
        "operationId": "admin_run_now",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "description": "`nudge_blockers` or `stalled_orders`",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobQueued"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error; see `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminStatus": {
        "type": "object",
        "required": [
          "environment",
          "email_mode",
          "automatic_email_enabled",
          "failed_jobs",
          "pending_jobs",
          "emails_needing_attention",
          "orders_missing_fx_rate"
        ],
        "properties": {
          "environment": {
            "type": "string"
          },
          "email_mode": {
            "type": "string",
            "description": "`dry_run` or `smtp`."
          },
          "smtp_host": {
            "type": [
              "string",
              "null"
            ]
          },
          "email_redirect_to": {
            "type": [
              "string",
              "null"
            ],
            "description": "Set during staging or the parallel run: all mail goes to this one address."
          },
          "automatic_email_enabled": {
            "type": "boolean"
          },
          "failed_jobs": {
            "type": "integer",
            "format": "int64"
          },
          "pending_jobs": {
            "type": "integer",
            "format": "int64"
          },
          "emails_needing_attention": {
            "type": "integer",
            "format": "int64"
          },
          "orders_missing_fx_rate": {
            "type": "integer",
            "format": "int64"
          },
          "latest_eur_rate_day": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `validation`, `stage_gate`, `currency_locked`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. For `validation` it names the field and the reason."
          }
        }
      },
      "FxFetch": {
        "type": "object",
        "required": [
          "from",
          "to"
        ],
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "Items_Job": {
        "type": "object",
        "description": "The list envelope: `{\"items\": [...]}`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "kind",
                "payload",
                "run_at",
                "attempts",
                "max_attempts",
                "created_at"
              ],
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int64"
                },
                "kind": {
                  "type": "string"
                },
                "payload": {
                  "type": "object"
                },
                "dedupe_key": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "run_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "attempts": {
                  "type": "integer",
                  "format": "int32"
                },
                "max_attempts": {
                  "type": "integer",
                  "format": "int32"
                },
                "locked_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "locked_by": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "completed_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "failed_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "last_error": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      },
...
      "JobQueued": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Null when an identical job was already queued."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "autocrm_session"
      }
    }
  },
  "security": [
    {
      "session_cookie": []
    },
    {
      "bearer": []
    }
  ],
...
}